requirements.txt
flask
beautifulsoup4>=4.5.3
icalendar>=3.11.3
mysql-connector==2.1.4
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
termcolor>=1.1.0